% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmode.R
\name{getmode}
\alias{getmode}
\title{The get mode function}
\usage{
getmode(v)
}
\arguments{
\item{v}{a vector of quantitative data}
}
\value{
the mode of the vector
}
\description{
The get mode function
}
\examples{
getmode(c(1, 2, 3, 3, 5))
}
