% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Makes a binomial distribution}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{number of samples}

\item{n}{size of sample}

\item{p}{probability of success}
}
\value{
table with distribution values
}
\description{
Makes a binomial distribution
}
\examples{
mybin(iter=1000, n=10, p=0.5)
}
