% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{A function for getting the mean confident intervals.}
\usage{
myci(sample)
}
\arguments{
\item{sample}{a sample}
}
\value{
confidence intervals of 95%
}
\description{
A function for getting the mean confident intervals.
}
\examples{
myci(rnorm(20, 14, 2))
}
