% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{mymaxlik}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{the likelihood function for a distribution}

\item{x}{the sample}

\item{param}{the parameter values to test for}

\item{...}{}
}
\value{
a graph and information about the parameter that maximizes the likelihood function
}
\description{
mymaxlik
}
\examples{
y = c(3, 3, 4, 3, 4, 5, 5, 4)
logbin = function(x, param) log(dbinom(x, prob=param, size=20))
mymaxlik(lfun=logbin, x=y, param=seq(0, 1, length=1000), main="Binomiial, n=20")
}
