% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{the mean of the distribution}

\item{sigma}{the standard devaition of the distribution}

\item{a}{the probability to find for the distribution}
}
\value{
A list with mu, sigma, and area
}
\description{
myncurve
}
\examples{
myncurve(5, 2, 2)
}
