% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{mypvalue}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{t_calc}

\item{xmax}{max x value on plots}

\item{n}{length of sample}

\item{alpha}{alpha value}
}
\value{
displays p-value areas and returns list with q and pvalue
}
\description{
mypvalue
}
\examples{
set.seed(55)
x1=rnorm(30,mean=25,sd=5)
t_calc = (mean(x1) - 24)/(sd(x1)/sqrt(length(x1)))

mypvalue(t_calc,n=30,alpha=0.05)

}
