% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of seats sold}

\item{gamma}{probability that the flightwill be overbooked}

\item{p}{probability of a show}
}
\value{
a list of tickets to be sold using discrete distribution,
         tickets to be sold using normal distribution, N, p, and gamma.
}
\description{
ntickets
}
\examples{
ntickets(400, 0.02, 0.95)
}
