% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.R
\name{pareto}
\alias{pareto}
\title{The Pareto Function}
\usage{
pareto(x, mn = "Pareto barplot", ...)
}
\arguments{
\item{x}{a replicant vector}

\item{mn}{the title of the pareto graph}

\item{...}{}
}
\value{
the pareto graph
}
\description{
The Pareto Function
}
\examples{
incidents=c(27,24,22,10)
MS=c("ED","PP","MO","TC")
Cause=rep(MS,incidents)
pareto(x=Cause)
}
