% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{adjmyboot}
\alias{adjmyboot}
\title{An adjusted bootstrap function}
\usage{
adjmyboot(iter = 10000, x, fun = "mean", alpha = 0.05, ...)
}
\arguments{
\item{iter}{number of samples to make}

\item{x}{sample to resample from}

\item{fun}{parameter to find}

\item{alpha}{confidence interval}

\item{...}{extra parameters}
}
\value{
a list of useful data containing the confident intervals, the function used, the value of x, and the value of x stat
}
\description{
An adjusted bootstrap function
}
\examples{
adjmyboot(x=rnorm(40, 1, 1))
}
